Move BasicMeasure files into models/measures/basic.*
