+++ include/fullscore/models/measures/base.h
