+++ src/actions/paste_measure_from_buffer_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/paste_measure_from_buffer_action.h>
 
-#include <fullscore/models/measure.h>
+#include <fullscore/models/measures/basic.h>
 #include <fullscore/models/note.h>
