+++ src/actions/yank_measure_to_buffer_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/yank_measure_to_buffer_action.h>
 
-#include <fullscore/models/measure.h>
+#include <fullscore/models/measures/basic.h>
 #include <fullscore/models/note.h>
