+++ src/factories/measure_grid_factory.cpp
@@ -2,8 +2,8 @@
 
 
 #include <fullscore/factories/measure_grid_factory.h>
+#include <fullscore/models/measures/basic.h>
 #include <fullscore/models/note.h>
-#include <fullscore/models/measure.h>
 #include <allegro_flare/useful.h>
 #include <iostream>
