+++ src/models/measures/base.cpp
@@ -1,7 +1,7 @@
 
 
 
-#include <fullscore/models/measure/base.h>
+#include <fullscore/models/measures/base.h>
